requirements.lock.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.9.2
requests==2.32.5
PyYAML==6.0

